feat(InputChecker.kt): add fun isWrongNumberInput, isWrongMoreOrStopCommand
